+++ minchin/pelican/plugins/cname.py
@@ -11,7 +11,7 @@ import os
 from pelican import signals
 
 __title__ = "minchin.pelican.plugins.cname"
-__version__ = "1.2.0"
+__version__ = "1.2.1-dev"
 __description__ = "Pelican plugin that adds a `CNAME` file to the output root. Useful for publishing to Github Pages. Written in Python."
 __author__ = "William Minchin"
 __email__ = "w_minchin@hotmail.com"
@@ -27,10 +27,13 @@ def add_cname(p):
 
     cname_path = os.path.join(p.output_path, 'CNAME')
     siteurl = p.settings.get('SITEURL', "127.0.0.1:8000")
-    siteurl = siteurl.replace("http://", "")
-    siteurl = siteurl.replace("https://", "")
-    if siteurl.startswith("//"):
+    if siteurl.startswith("http://"):
+        siteurl = siteurl[7:]
+    elif siteurl.startswith("https://"):
+        siteurl = siteurl[8:]
+    elif siteurl.startswith("//"):
         siteurl = siteurl[2:]
+
     with open(cname_path, 'w') as cname_file:
         cname_file.write(siteurl)