+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ CName Changelog
 
 This Changelog is for ``minchin.pelican.plugins.cname``.
 
+- :bug:`-` only replace protocol at the beginning of SITEURL's
 - :release:`1.2.0 <2018-10-25>`
 - :feature:`-` deal with protocol-less SITEURL's
 - :release:`1.1.0 <2018-10-25>`